breaking change: tidy up all columns in Columns/ folder and \Columns Namespace
