+++ src/Columns/BooleanColumn.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mediconesystems\LivewireDatatables;
+namespace Mediconesystems\LivewireDatatables\Columns;
 
 class BooleanColumn extends Column
 {
