+++ src/Columns/JsonColumn.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Mediconesystems\LivewireDatatables;
+namespace Mediconesystems\LivewireDatatables\Columns;
 
 class JsonColumn extends Column
 {
