+++ src/Http/Livewire/LivewireDatatable.php
@@ -17,8 +17,8 @@ use Illuminate\View\View;
 use Livewire\Component;
 use Livewire\WithPagination;
 use Maatwebsite\Excel\Facades\Excel;
-use Mediconesystems\LivewireDatatables\Column;
-use Mediconesystems\LivewireDatatables\ColumnSet;
+use Mediconesystems\LivewireDatatables\Columns\Column;
+use Mediconesystems\LivewireDatatables\Columns\ColumnSet;
 use Mediconesystems\LivewireDatatables\Exports\DatatableExport;
 use Mediconesystems\LivewireDatatables\Traits\WithCallbacks;
 use Mediconesystems\LivewireDatatables\Traits\WithPresetDateFilters;
